In Development

OrbitalTrajectories.jl and its documentation is currently in development and is provided as a pre-1.0 release.

The interfaces (exported types/functions) here may, and most likely will, change in future updates, possibly breaking any existing code using OrbitalTrajectories.jl.

While these interfaces are being developed, the documentation provided here is of very limited use. For now, please instead refer to the tutorials/examples which will be kept up-to-date as best as possible.

Core.Type — Method

Generic constructor for a DynamicalModel's underlying ODEFunctions.

OrbitalTrajectories.Dynamics.Abstract_DynamicalModel — Method

Generic model function.

OrbitalTrajectories.Dynamics.STM_ODEFunction — Method

Function to generate an ODE Function that computes the State Transition
Matrix simultaneously with the given function f. This requires f to have
a computable Jacobian function.

OrbitalTrajectories.Dynamics.centrifugal_potential — Method

Centrifugal potential [DeiTos2017, Eq.13; Ichinomiya 2018, Eq. 2.2]

OrbitalTrajectories.Dynamics.elliptical_potential — Method

Elliptical potential in barycentric frame [DeiTos2017, Eq. 15]

OrbitalTrajectories.Dynamics.hill_sphere_radius — Method

Compute the Hill sphere radius for two bodies

OrbitalTrajectories.Dynamics.jacobi — Method

Jacobi integral = 2U - V^2 = -2E [Oshima 2019, Eq. 3; Koon, Eq. 2.3.13; Wakker, Eq. 3.54]

OrbitalTrajectories.Dynamics.pos_to_inertial — Method

Returns the position rotation matrix from a rotating 3BP frame to an inertial frame.

OrbitalTrajectories.Dynamics.sensitivity — Method

Sensitivity of the final state with respect to initial state.

OrbitalTrajectories.Dynamics.sensitivity_trace — Method

Sensitivity of the full propagated trajectory with respect to initial state.

OrbitalTrajectories.Dynamics.state_to_inertial — Method

Returns the state rotation matrix from a rotating 3BP frame to an inertial frame.

OrbitalTrajectories.SpiceUtils.dynamic_synodic_frame — Method

Automatically build and return a dynamic frame for a synodic reference (b1 is the observer, b2 is the target).

OrbitalTrajectories.SpiceUtils.get_GM — Method

Get body gravitational constant (GM) from SPICE kernels

OrbitalTrajectories.SpiceUtils.get_pos — Method

Get a target body position from SPICE kernels.

OrbitalTrajectories.SpiceUtils.get_state — Method

Get a target body state from SPICE kernels

OrbitalTrajectories.SpiceUtils.mass_fraction — Method

Get the mass fraction of two bodies (smaller mass divided by total mass)